pyproject.toml
[tool.poetry]
name = "toggl-report-bq"
version = "0.1.0"
description = ""
authors = ["kawadone121 <ryoto.kawahara@gmail.com>"]
readme = "README.md"
packages = [{include = "toggl_report_bq"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
pandas = "^1.5.1"
google-cloud-bigquery = "^3.3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
